requirements.txt
Flask
gunicorn
numpy
opencv-contrib-python-headless
tensorflow-cpu
keras
scikit-learn
scipy
seaborn
requests
joblib
Jinja2
itsdangerous
MarkupSafe
Werkzeug
jsonify